Cleanup: move bmqsys_statmonitorsnapshotrecorder -> mqbstat_statmonitorsnapshotrecorder
