+++ src/groups/bmq/bmqsys/doc/bmqsys.txt
@@ -9,11 +9,9 @@ components to be reused through various applications.
 
 /Hierarchical Synopsis
 /---------------------
-The 'bmqsys' package currently has 2 components having 2 levels of physical
+The 'bmqsys' package currently has 1 component having 1 level of physical
 dependency.  The list below shows the hierarchal ordering of the components.
 ..
-  2. bmqsys_statmonitorsnapshotrecorder
-
   1. bmqsys_statmonitor
 ..
 
@@ -21,6 +19,3 @@ dependency.  The list below shows the hierarchal ordering of the components.
 /------------------
 : 'bmqsys_statmonitor':
 :      Provide a mechanism to monitor cpu and memory stats.
-:
-: 'bmqsys_statmonitorsnapshotrecorder':
-:      Provide a mechanism to record snapshot of time, cpu, and ctx switch.
