+++ src/groups/bmq/bmqsys/package/bmqsys.mem
@@ -1,2 +1,1 @@
 bmqsys_statmonitor
-bmqsys_statmonitorsnapshotrecorder
