+++ src/groups/mqb/mqbstat/package/mqbstat.mem
@@ -6,3 +6,4 @@ mqbstat_jsonprinter
 mqbstat_printer
 mqbstat_queuestats
 mqbstat_statcontroller
+mqbstat_statmonitorsnapshotrecorder
